Refactor CardTravelInfoPassenger styles
