+++ components/Cards/CardTravelInfoPassenger.jsx
@@ -1,4 +1,5 @@
-import { StyleSheet, Text, View } from "react-native";
+import { Text, View } from "react-native";
+import styles from "@/styles/CardTravelInfoPassenger";
 
 export default function CardTravelInfoPassenger() {
   return (
@@ -52,46 +53,3 @@ export default function CardTravelInfoPassenger() {
   );
 }
 
-const styles = StyleSheet.create({
-  cardContainer: {
-    width: 335,
-    height: 273,
-    borderWidth: 2,
-    borderColor: "#40498D",
-    borderRadius: 20,
-    paddingHorizontal: 25,
-  },
-  driverInfoContainer: {
-    marginTop: 18,
-    gap: 18,
-  },
-  infoLabel: {
-    fontSize: 19,
-    fontWeight: "700",
-    color: "#40498D",
-  },
-  infoText: {
-    fontSize: 16,
-    fontWeight: "700",
-  },
-  rowContainer: {
-    flexDirection: "row",
-    gap: 100,
-    marginTop: 10,
-  },
-  seatsContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-  },
-  seatView: {
-    width: 20,
-    height: 20,
-    backgroundColor: "#40498D",
-    alignItems: "center",
-    borderRadius: 6,
-  },
-  seatText: {
-    color: "white",
-    fontWeight: "700",
-  },
-});
